Se aniade soporte a id de regla mayor a 9 Cooreccion en programacion extremisima D:

+++ src/se/determinista/arbol/Arbol.java
@@ -16,7 +16,7 @@ public class Arbol
             arbol = nodo;
     }
 
-    public long getReglaDirLogica(byte llave) {
+    public long getReglaDirLogica(Integer llave) {
         return buscarReglaDirLogica(llave, arbol);
     }
